Support excluding validation attribute values
The optional exclude attribute for Validators allows unique attributes
that are not changing to be included in patch requests. Originally,
these fields would fail validation as they are not unique, but the
current value can now be excluded in the validation.

Updated accounts variable in the err_if_not_found decorator to
something more generic and updated the backref on the User model to
show the user and not school.