Fix scrollbar and fixed bug on mobile
On mobile, the scrollbar was always showing on all pages, even when
there was no more content to show. Now the scrollbar only shows if
additional content is present. Additionally, position fixed messes up
scrolling after margins on mobile so position was changed to fixed for
the ViewAccount sticky buttons.