Add support for running application in Docker
The application can now be developed using docker-compose and deployed
in production through docker-containers.  To support this nginx was
implemented as a proxy server to route all requests to the react
application.

The backend and the fronted both have individual Dockerfiles that allow
them to be deployed individually in production.