Add a 404 page for not found requested resources
Users will receive a 404 page if the resource they requested does not
exist. Routes requesting specific resources have been updated to render
the 404 error page if the resource could not be found. Additionally, all
invalid routes will also be redirected to the page even if no specific
resource is requested.